refactor: エンジンをモジュール分割し、bool引数をenum化
1347行の conversion.rs を chunk.rs / strategy.rs と同じ兄弟モジュール方式で分割:
- model.rs — モデルディスパッチとその前段のキャッシュ、ウィンドウ変換
- filter.rs — Ctrl+R/Ctrl+T のソース絞り込みビュー一式
- conversion.rs — 混合候補リストの構築・キー処理・確定（854行）

あわせて呼び出し側の可読性を上げる整理:
- start_conversion/build_conversion_candidates の bool を LearningLookup に
- cycle_candidate_filter/start_filtered_conversion の bool を FilterDirection に
- InputState::reading() アクセサを追加し、状態の分解 match を整理
- 呼び手のいない CandidateList::reset/update と Candidate の From impl を削除
- cancel_conversion を1行で包むだけの backspace_conversion を削除
