+++ karukan-im/core/src/core/state.rs
@@ -63,6 +63,14 @@ impl InputState {
         }
     }
 
+    /// The reading a conversion was built from, if in the Conversion state.
+    pub fn reading(&self) -> Option<&str> {
+        match self {
+            Self::Conversion { reading, .. } => Some(reading),
+            _ => None,
+        }
+    }
+
     /// Get candidates in conversion state
     pub fn candidates(&self) -> Option<&CandidateList> {
         match self {
